+++ lib/include/database/table_reader.hpp
@@ -25,7 +25,7 @@ public:
 
   template < class T >
   void read_value(const char *, T &) {}
-  void read_value(const char *, char *, int ) {}
+  void read_value(const char *, char *, size_t) {}
   void read_value(const char *, object_base_ptr &x);
   void read_value(const char *id, object_container &x);
