+++ lib/include/object/object_deleter.hpp
@@ -116,7 +116,7 @@ public:
   template < class T >
   void read_value(const char*, const T&) {}
 
-  void read_value(const char*, char*, int) {}
+  void read_value(const char*, char*, size_t) {}
   void read_value(const char*, object_base_ptr &x);
   void read_value(const char*, object_container &x);
   void read_value(const char*, basic_identifier &x);
