+++ lib/src/database/table.cpp
@@ -49,7 +49,7 @@ public:
   template < class T >
   void read_value(const char*, T&) {}
 
-  void read_value(const char*, char*, int) {}
+  void read_value(const char*, char*, size_t) {}
 
   void read_value(const char *id, object_base_ptr &x)
   {
